feat: restructure blockchain and wallet services, enhance error handling, and improve response utilities
