+++ src/features/wallet/services/walletServices.ts
@@ -11,12 +11,21 @@ export type WalletKeys = {
   publicKey: string
   address: string
 }
+
 function serializeTransaction(tx: Pick<Transaction, "from" | "to" | "amount" | "timestamp">): string {
   return `${tx.from}|${tx.to}|${tx.amount}|${tx.timestamp}`
 }
+
 export const walletService = {
-  generateWallet: async (): Promise<WalletKeys> => {
-    const mnemonic = bip39.generateMnemonic()
+  generateWallet: async (wordCount: number = 12): Promise<WalletKeys> => {
+    const validWordCounts = [12, 15, 18, 21, 24]
+    if (!validWordCounts.includes(wordCount)) {
+      throw new Error(`Invalid mnemonic word count. Must be one of: ${validWordCounts.join(", ")}`)
+    }
+    const entropyBits = 32 * (wordCount / 3)
+    const entropyBytes = entropyBits / 8
+    const entropyBuffer = crypto.randomBytes(entropyBytes)
+    const mnemonic = bip39.entropyToMnemonic(entropyBuffer.toString("hex"))
     const seed = await bip39.mnemonicToSeed(mnemonic)
 
     const keyPair = ec.keyFromPrivate(seed.toString("hex").slice(0, 64))
@@ -53,7 +62,6 @@ export const walletService = {
   },
 
   signTransaction: (transaction: Transaction, privateKey: string): string => {
-    // Only take the data fields, not signature or publicKey
     const payload = serializeTransaction(transaction)
     const key = ec.keyFromPrivate(privateKey, "hex")
     const hash = crypto.createHash("sha256").update(payload).digest()
